# common/api/content/v1/tag_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.33.1
// source: content/v1/tag.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationContentTagServiceAdds = "/common.api.content.v1.ContentTagService/Adds"
const OperationContentTagServicePage = "/common.api.content.v1.ContentTagService/Page"
const OperationContentTagServiceUpdate = "/common.api.content.v1.ContentTagService/Update"

type ContentTagServiceHTTPServer interface {
	// Adds 批量添加标签
	Adds(context.Context, *AddTagsRequest) (*AddTagsReply, error)
	// Page 分页获取标签
	Page(context.Context, *PageTagRequest) (*PageTagReply, error)
	// Update 更新标签
	Update(context.Context, *UpdateTagRequest) (*UpdateTagReply, error)
}

func RegisterContentTagServiceHTTPServer(s *http.Server, srv ContentTagServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/tag/adds", _ContentTagService_Adds0_HTTP_Handler(srv))
	r.POST("/v1/tag/update", _ContentTagService_Update0_HTTP_Handler(srv))
	r.POST("/v1/tag/page", _ContentTagService_Page0_HTTP_Handler(srv))
}

func _ContentTagService_Adds0_HTTP_Handler(srv ContentTagServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddTagsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentTagServiceAdds)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Adds(ctx, req.(*AddTagsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddTagsReply)
		return ctx.Result(200, reply)
	}
}

func _ContentTagService_Update0_HTTP_Handler(srv ContentTagServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTagRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentTagServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*UpdateTagRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTagReply)
		return ctx.Result(200, reply)
	}
}

func _ContentTagService_Page0_HTTP_Handler(srv ContentTagServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageTagRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentTagServicePage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Page(ctx, req.(*PageTagRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageTagReply)
		return ctx.Result(200, reply)
	}
}

type ContentTagServiceHTTPClient interface {
	// Adds 批量添加标签
	Adds(ctx context.Context, req *AddTagsRequest, opts ...http.CallOption) (rsp *AddTagsReply, err error)
	// Page 分页获取标签
	Page(ctx context.Context, req *PageTagRequest, opts ...http.CallOption) (rsp *PageTagReply, err error)
	// Update 更新标签
	Update(ctx context.Context, req *UpdateTagRequest, opts ...http.CallOption) (rsp *UpdateTagReply, err error)
}

type ContentTagServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewContentTagServiceHTTPClient(client *http.Client) ContentTagServiceHTTPClient {
	return &ContentTagServiceHTTPClientImpl{client}
}

// Adds 批量添加标签
func (c *ContentTagServiceHTTPClientImpl) Adds(ctx context.Context, in *AddTagsRequest, opts ...http.CallOption) (*AddTagsReply, error) {
	var out AddTagsReply
	pattern := "/v1/tag/adds"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentTagServiceAdds))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Page 分页获取标签
func (c *ContentTagServiceHTTPClientImpl) Page(ctx context.Context, in *PageTagRequest, opts ...http.CallOption) (*PageTagReply, error) {
	var out PageTagReply
	pattern := "/v1/tag/page"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentTagServicePage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新标签
func (c *ContentTagServiceHTTPClientImpl) Update(ctx context.Context, in *UpdateTagRequest, opts ...http.CallOption) (*UpdateTagReply, error) {
	var out UpdateTagReply
	pattern := "/v1/tag/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentTagServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}