# common/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/tag/adds:
        post:
            tags:
                - ContentTagService
            description: 批量添加标签
            operationId: ContentTagService_Adds
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AddTagsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AddTagsReply'
    /v1/tag/page:
        post:
            tags:
                - ContentTagService
            description: 分页获取标签
            operationId: ContentTagService_Page
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.PageTagRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.PageTagReply'
    /v1/tag/update:
        post:
            tags:
                - ContentTagService
            description: 更新标签
            operationId: ContentTagService_Update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.UpdateTagRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.UpdateTagReply'
...
components:
    schemas:
...
        common.api.common.v1.Int32Range:
            type: object
            properties:
                start:
                    type: integer
                    description: 开始
                    format: int32
                end:
                    type: integer
                    description: 结束
                    format: int32
        common.api.common.v1.PageReply:
            type: object
            properties:
                total:
                    type: integer
                    description: 总数
                    format: uint32
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
        common.api.common.v1.PageRequest:
            type: object
            properties:
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
...
        common.api.content.v1.AddTagsReply:
            type: object
            properties:
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Tag'
                    description: 添加的标签信息
        common.api.content.v1.AddTagsRequest:
            type: object
            properties:
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.TagSave'
                    description: 标签信息
...
        common.api.content.v1.PageTagReply:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageReply'
                    description: 分页
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Tag'
                    description: 标签列表
        common.api.content.v1.PageTagRequest:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageRequest'
                    description: 分页
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.TagQueryParams'
                    description: 查询参数
...
        common.api.content.v1.Tag:
            type: object
            properties:
                created_at:
                    type: string
                    description: 创建时间
                    format: date-time
                updated_at:
                    type: string
                    description: 更新时间
                    format: date-time
                created_by:
                    type: string
                    description: 创建人
                updated_by:
                    type: string
                    description: 更新人
                id:
                    type: string
                    description: 主键
                name:
                    type: string
                    description: 标签名称
                description:
                    type: string
                    description: 描述
                domain_id:
                    type: string
                    description: 所属领域Id
                status:
                    type: integer
                    description: 标签状态：0-正常，1-禁用
                    format: int32
                article_count:
                    type: integer
                    description: 文章数
                    format: int32
        common.api.content.v1.TagQueryParams:
            type: object
            properties:
                ids:
                    type: array
                    items:
                        type: string
                    description: 标签Id列表
                user_id:
                    type: string
                    description: 用户Id
                name:
                    type: string
                    description: 标签名称
                names:
                    type: array
                    items:
                        type: string
                    description: 标签名称列表
                description:
                    type: string
                    description: 标签描述
                status:
                    type: integer
                    description: 标签状态
                    format: int32
                domain_id:
                    type: string
                    description: 领域Id
                article_count:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.Int32Range'
                    description: 文章数量范围
        common.api.content.v1.TagSave:
            type: object
            properties:
                id:
                    type: string
                    description: 主键
                domain_id:
                    type: string
                    description: 领域Id
                status:
                    type: integer
                    description: 状态：0-正常，1-禁用
                    format: int32
                name:
                    type: string
                    description: 标签名称
                description:
                    type: string
                    description: 标签描述
...
        common.api.content.v1.UpdateTagReply:
            type: object
            properties:
                tag:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.Tag'
                    description: 更新的标签信息
        common.api.content.v1.UpdateTagRequest:
            type: object
            properties:
                tag:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.TagSave'
                    description: 标签信息
...